% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_cols.R
\name{all_nmfs_palettes}
\alias{all_nmfs_palettes}
\title{Return function that shows all nmfs color palettes}
\usage{
all_nmfs_palettes()
}
\description{
Return function that shows all nmfs color palettes
}
\examples{
all_nmfs_palettes()
}
