% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_cols.R
\name{display_nmfs_palette}
\alias{display_nmfs_palette}
\title{Return interpolated nmfs color palette}
\usage{
display_nmfs_palette(name, n)
}
\arguments{
\item{name}{Character name of palette in nmfs_palettes.}

\item{n}{Number of colors in palette.}
}
\value{
A list object showing a specific nmfs color palette in the plot window.
}
\description{
Return interpolated nmfs color palette
}
\examples{
display_nmfs_palette("oceans", 10)
}
