% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_cols.R
\name{nmfs_cols}
\alias{nmfs_cols}
\title{Function to extract nmfs colors as hex codes}
\usage{
nmfs_cols(...)
}
\arguments{
\item{...}{Character names of nmfs_colors}
}
\description{
Function to extract nmfs colors as hex codes
}
\examples{
nmfs_cols("processblue")
}
