% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_cols.R
\name{nmfs_palette}
\alias{nmfs_palette}
\title{Return function to interpolate a nmfs color palette}
\usage{
nmfs_palette(palette = "oceans", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in nmfs_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a nmfs color palette
}
\examples{
nmfs_palette("oceans")(10)
}
