% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfspalette-package.R
\docType{package}
\name{nmfspalette-package}
\alias{nmfspalette}
\alias{nmfspalette-package}
\title{nmfspalette: A package for NOAA Fisheries color schemes}
\description{
The nmfspalette package provides functions to extract colors, color
palettes, and use ggplot2 palettes in R.
}
\section{Details}{

This package contains functions to support graphing in base R and ggplot2
using NOAA Fisheries colors.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nmfs-ost/nmfspalette}
  \item Report bugs at \url{https://github.com/nmfs-ost/nmfspalette/issues}
}

}
\author{
\strong{Maintainer}: Sophie Breitbart \email{sophie.breitbart@noaa.gov}

Authors:
\itemize{
  \item Christine Stawitz \email{christine.stawitz@noaa.gov}
  \item Bai Li \email{bai.li@noaa.gov}
}

Other contributors:
\itemize{
  \item Kathryn Doering \email{kathryn.doering@noaa.gov} [contributor]
  \item Steve Saul \email{steven.saul@noaa.gov} [contributor]
  \item Sam Schiano \email{samantha.schiano@noaa.gov} [contributor]
}

}
\keyword{internal}
