% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfspalette.R
\docType{package}
\name{nmfspalette}
\alias{nmfspalette}
\title{nmfspalette: A package for NOAA Fisheries color schemes}
\description{
The nmfspalette package provides functions to extract colors, color palettes, and use ggplot2
palettes in R.
}
