% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_cols.R
\name{display_nmfs_palette}
\alias{display_nmfs_palette}
\title{Return function to interpolate a nmfs color palette}
\usage{
display_nmfs_palette(name, n, ...)
}
\arguments{
\item{name}{Character name of palette in nmfs_palettes.}

\item{n}{Number of colors in palette.}

\item{...}{Additional arguments to pass to \code{\link[graphics:image]{graphics::image()}}.}
}
\description{
Return function to interpolate a nmfs color palette
}
\examples{
display_nmfs_palette("oceans", 10)
}
