% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{scale_color_nmfs}
\alias{scale_color_nmfs}
\title{Color scale constructor for nmfs colors}
\usage{
scale_color_nmfs(palette = "oceans", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{nmfs_palettes}. Default value
is "oceans".}

\item{discrete}{Boolean indicating whether color aesthetic is discrete.
Default is TRUE.}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default is FALSE.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}, used respectively when
\code{discrete} is TRUE or FALSE.}
}
\description{
Color scale constructor for nmfs colors
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_nmfs("coral")
}
