% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_cols.R
\name{scale_color_nmfs}
\alias{scale_color_nmfs}
\title{Color scale constructor for nmfs colors}
\usage{
scale_color_nmfs(palette = "oceans", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in nmfs_palettes.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to \code{discrete_scale()} or
\code{scale_color_gradientn()}, used respectively when discrete is
TRUE or FALSE.}
}
\description{
Color scale constructor for nmfs colors
}
\examples{
\dontrun{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_nmfs("coral")
}
}
