% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{scale_fill_nmfs}
\alias{scale_fill_nmfs}
\title{Fill scale constructor for nmfs colors}
\usage{
scale_fill_nmfs(palette = "oceans", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{nmfs_palettes}. Default value
is "oceans".}

\item{discrete}{Boolean indicating whether color aesthetic is discrete.
Default value is TRUE.}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default value is FALSE.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}, used respectively when
\code{discrete} is TRUE or FALSE.}
}
\description{
Fill scale constructor for nmfs colors
}
\examples{
\dontrun{
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_nmfs(palette = "crustacean", discrete = FALSE)
}
}
