% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_cols.R
\name{scale_fill_nmfs}
\alias{scale_fill_nmfs}
\title{Fill scale constructor for nmfs colors}
\usage{
scale_fill_nmfs(palette = "oceans", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in nmfs_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for nmfs colors
}
\examples{
\dontrun{
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_nmfs(palette = "crustacean", discrete = FALSE)
}
}
