% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{theme_nmfs}
\alias{theme_nmfs}
\title{Create theme for nmfs colors}
\usage{
theme_nmfs(
  palette = "oceans",
  discrete = TRUE,
  reverse = FALSE,
  interpolate = TRUE,
  base_size = 14,
  lab_size = 12,
  ink = "black",
  paper = "white",
  accent = "#003087",
  ...
)
}
\arguments{
\item{palette}{Character name of palette in \code{nmfs_palettes}. Default value
is "oceans".}

\item{discrete}{Boolean indicating whether color aesthetic is discrete.
Default is TRUE.}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default is FALSE.}

\item{interpolate}{Boolean indicating whether the colors assigned to plot
objects should interpolated from palettes, with the alternative that only the
defined colors in the palette are used. Default is TRUE.}

\item{base_size}{The base font size, as defined in the
\code{\link[ggplot2]{theme_gray}} function}

\item{lab_size}{The axis text size}

\item{ink}{The plot foreground color, as defined in the
\code{\link[ggplot2]{theme_gray}} function}

\item{paper}{The plot background color, as defined in the
\code{\link[ggplot2]{theme_gray}} function}

\item{accent}{The plot's accented elements' color, as defined in the
\code{\link[ggplot2]{theme_gray}} function}

\item{...}{Additional arguments passed to: \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}
when \code{discrete} is TRUE; \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} when \code{discrete} is FALSE
and \code{interpolate} is TRUE; and \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}} when \code{discrete}
is FALSE and \code{interpolate} is FALSE.}
}
\value{
A theme that applies nmfs color palettes and theme-related elements (such as label sizes, borders, axis lines, and more) to a plot.
}
\description{
Create theme for nmfs colors
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = disp, color = as.factor(cyl))) +
  geom_point(size = 3) +
  theme_nmfs(discrete = TRUE, interpolate = TRUE, palette = "crustacean")
ggplot(mtcars, aes(x = mpg, y = disp, color = as.factor(cyl))) +
  geom_point(size = 3) +
  theme_nmfs(discrete = TRUE, interpolate = FALSE, palette = "crustacean")
ggplot(mtcars, aes(x = mpg, y = disp, fill = hp)) +
geom_point(size = 3, shape = 24) +
theme_nmfs(discrete = FALSE, interpolate = TRUE, palette = "seagrass")
}
